Multiple components: Implement bare-bones hover support
This implementation looks up hovered symbols in the type-data file that
games/mods should check in to source control at `<root>/.oraide/type-data.json`
which can be generated via the current work-in-progress and not upstream
utility command `--type-data`.

See https://github.com/Phrohdoh/OpenRA/tree/oraide-util-extract-type-data
for the time being.  Ideally this would be merged upstream at which point the
branch would be deleted causing the previous link to become a 404.