Implement the `SdkCtx` query group
This gives the query system access to `oraide_sdk::Game`s as long as a
workspace root has been set prior to calling `SdkCtx::all_games`.

What this means is that we can eventually questions like:
- "What is the title of game X?"
- "What are all of the file paths defined in the `Rules` section of game X's manifest?"
Where 'X' is a placeholder for a Game's `id`.

In this implementation `SdkCtx` requires the implementor to also implement
`LanguageServerCtx` only because `workspace_root` is part of
`LanguageServerCtx` which should be changed.

oraide-cli:
- Add a `list-games` command

oraide-query-system:
- Add `SdkCtxStorage` to `OraideDatabase`

oraide-sdk:
- Delete every thing that previously existed
- Add type `GameId` which is a newtype over `std`'s `String`
- Add a `Game` type with an `id` field of type `GameId`
- Define and implement `SdkCtx`
- Export `SdkCtx`, `SdkCtxStorage`, `Game`, and `GameId` from the crate root